Moving generated types to their own file
My script now produces its own file to make it easier to quickly generate the types
